requirements.txt
kafka-python
fastapi
uvicorn
pandas
pyspark
pydantic
scikit-learn

